[Silver III] Title: 주유소, Time: 388 ms, Memory: 36116 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -1,7 +1,9 @@
+// Don't place your source in a package
 import java.util.*;
 import java.lang.*;
 import java.io.*;
 
+// Please name your class Main
 class Main {
 	public static void main (String[] args) throws IOException {
 	    BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
@@ -9,19 +11,19 @@ class Main {
 	    Integer N = Integer.parseInt(br.readLine());  // 도시 개수
 	    
 	    StringTokenizer st = new StringTokenizer(br.readLine());
-	    int[] cityLen = new int[N - 1];  // 도시 사이 길이
+	    long[] cityLen = new long[N - 1];  // 도시 사이 길이
 	    for (int i = 0; i < N - 1; i++) {
-	        cityLen[i] = Integer.parseInt(st.nextToken());
+	        cityLen[i] = Long.parseLong(st.nextToken());
 	    }
 	    
 	    st = new StringTokenizer(br.readLine());
-	    int[] pricePerLiter = new int[N];  // 리터 당 가격
+	    long[] pricePerLiter = new long[N];  // 리터 당 가격
 	    for (int i = 0; i < N - 1; i++) {
-	        pricePerLiter[i] = Integer.parseInt(st.nextToken());
+	        pricePerLiter[i] = Long.parseLong(st.nextToken());
 	    }
 	    
-	    int currMin = pricePerLiter[0];
-	    int answer = 0;
+	    long currMin = pricePerLiter[0];
+	    long answer = 0;
 	    
 	    answer += currMin * cityLen[0];
